refactors react-discovery-solr adds dynamic block join query builder
