+++ packages/react-discovery-solr/src/query-builders/FacetTypes.ts
