+++ packages/react-discovery-solr/src/query-builders/SolrParameters.ts
