+++ packages/react-discovery-solr/src/query-builders/__test__/es-response.json
