+++ packages/react-discovery-solr/src/query-builders/__test__/solr-response.json
