+++ packages/react-discovery-solr/src/query-builders/index.ts
@@ -1,0 +1,5 @@
+export * from '../components'
+export * from './queryBuilder'
+export * from './SolrParameters'
+export * from './suggestQueryBuilder'
+
