+++ packages/react-discovery-solr/src/query-builders/suggestQueryBuilder.ts
